Update testsAuto structure
